move(models): from domain to business
