+++ src/business/calculations/coordinates.py
@@ -2,10 +2,10 @@ import string
 from decimal import Decimal as _
 
 import errors
-from domain.models import Coordinate, CoordinatePair, Nomenclature
 from misc import from_tuple
 
 from business.math_actions import coordinate_actions, coordinate_pair_actions
+from business.models import Coordinate, CoordinatePair, Nomenclature
 from business.types import NomenclatureTitleFormatter
