+++ src/domain/facades.py
@@ -1,8 +1,7 @@
 import abc
 
+from business.models import CoordinatePair
 from business.types import ImageType
-
-from domain.models import CoordinatePair
 
 
 class INomenclatureFacade(abc.ABC):
