+++ tests/test_chain_by_nomenclature_title.py
@@ -1,7 +1,7 @@
 import pytest
 from business.enums import Scale
+from business.models import CoordinatePair, Nomenclature
 from business.scale_resolver import nomenclature_title_resolver
-from domain.models import CoordinatePair, Nomenclature
 from misc import from_tuple
 
 from .test_helper import generate_random_bound
